backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
langgraph==0.2.53
sentence-transformers==3.3.0
faiss-cpu==1.9.0
PyMuPDF==1.24.14
python-docx==1.1.2
tavily-python==0.5.0
pydantic==2.9.2
